Add Projects [updateProc]

+++ backend/projects/HESS/src/org/mdoubleh/www/action/DeleteAction.java
@@ -10,9 +10,9 @@ import org.mdoubleh.www.common.ActionForward;
 import org.mdoubleh.www.common.LoginManager;
 import org.mdoubleh.www.common.RegExp;
 import org.mdoubleh.www.service.HessService;
-import org.mdoubleh.www.vo.ArticleVo;
 
-import static org.mdoubleh.www.common.RegExp.ARTICLE_NUM;;
+import static org.mdoubleh.www.common.RegExp.ARTICLE_NUM;
+import static org.mdoubleh.www.common.RegExp.IS_NUMBER;
 
 public class DeleteAction implements Action {
 	@Override
@@ -20,27 +20,27 @@ public class DeleteAction implements Action {
 		LoginManager lm = LoginManager.getInstance();
 		String id = lm.getMemberId(request.getSession());
 		if (id == null) {
-			response.setContentType("text/html; charset=UTF-8");
+			response.setContentType("text/html;charset=UTF-8");
 			PrintWriter out = response.getWriter();
-			out.println("<script>alert('로그인이 필요한 서비스 입니다.'); location.href='/login.do';</script>");
+			out.println("<script>alert('로그인이 필요한 서비스 입니다.');location.href='/login.do';</script>");
 			out.close();
 			return null;
 		}
 		
 		String num = request.getParameter("num");
 		if (num == null || num.equals("") || !RegExp.checkString(ARTICLE_NUM, num)) {
-			response.setContentType("text/html; charset=UTF-8");
+			response.setContentType("text/html;charset=UTF-8");
 			PrintWriter out = response.getWriter();
-			out.println("<script>alert('잘못된 접근입니다.'); location.href='/index.jsp';</script>");
+			out.println("<script>alert('잘못된 접근입니다.');location.href='/';</script>");
 			out.close();
 			return null;
 		}
 		
 		int buff = Integer.parseInt(num);
 		if (buff <= 0) {
-			response.setContentType("text/html; charset=UTF-8");
+			response.setContentType("text/html;charset=UTF-8");
 			PrintWriter out = response.getWriter();
-			out.println("<script>alert('잘못된 접근입니다.'); location.href='/index.jsp';</script>");
+			out.println("<script>alert('잘못된 접근입니다.');location.href='/';</script>");
 			out.close();
 			return null;
 		}
@@ -48,16 +48,43 @@ public class DeleteAction implements Action {
 		HessService service = new HessService();
 		
 		String writerId = service.getWriterId(buff);
-		if (!service.deleteArticle(buff)) {
-			response.setContentType("text/html; charset=UTF-8");
+		if (writerId == null || !id.equals(writerId)) {
+			response.setContentType("text/html;charset=UTF-8");
 			PrintWriter out = response.getWriter();
-			out.println("<script>alert('글을 삭제하는데 실패하였습니다.'); location.href='/';</script>");
+			out.println("<script>alert('잘못된 접근입니다.');location.href='/';</script>");
+			out.close();
+			return null;
+		}
+		
+		if (!service.deleteArticle(buff)) {
+			response.setContentType("text/html;charset=UTF-8");
+			PrintWriter out = response.getWriter();
+			out.println("<script>alert('글을 삭제하는데 실패하였습니다.');location.href='/';</script>");
+			out.close();
+			return null;
+		}
+		
+		String pageNum = request.getParameter("pn");
+		if (pageNum == null
+				|| !RegExp.checkString(IS_NUMBER, pageNum)) {
+			response.setContentType("text/html;charset=UTF-8");
+			PrintWriter out = response.getWriter();
+			out.println("<script>alert('잘못된 접근입니다.');location.href='/';</script>");
+			out.close();
+			return null;
+		}
+		
+		int page = Integer.parseInt(pageNum);
+		if (page < 1) {
+			response.setContentType("text/html;charset=UTF-8");
+			PrintWriter out = response.getWriter();
+			out.println("<script>alert('잘못된 접근입니다.');location.href='/';</script>");
 			out.close();
 			return null;
 		}
 		
 		ActionForward forward = new ActionForward();
-		forward.setPath("/list.do");
+		forward.setPath("/list.do?pn=" + page);
 		forward.setRedirect(true);
 		return forward;
 	}
